+++ src/main/java/com/smartlogis/gatewayserver/GatewayServerApplication.java
@@ -2,7 +2,9 @@ package com.smartlogis.gatewayserver;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 
+@EnableFeignClients
 @SpringBootApplication
 public class GatewayServerApplication {
