+++ src/main/java/com/smartlogis/gatewayserver/security/KeycloakClientRoleConverter.java
@@ -1,4 +1,4 @@
-package com.smartlogis.gatewayserver.config;
+package com.smartlogis.gatewayserver.security;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.security.core.GrantedAuthority;
